+++ client/components/app-layout/Menu.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from 'next-i18next'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
-import React from 'react'
-import { cn } from 'simple-react-ui-kit'
+import React, { useState } from 'react'
+import { Icon, cn } from 'simple-react-ui-kit'
 
 import styles from './styles.module.sass'
 
@@ -10,16 +10,24 @@ export type MenuItemType = {
     link: string
     text: string
     external?: boolean
+    subMenuItems?: MenuItemType[]
 }
 
 interface MenuProps {
     className?: string
+    sidebarMenu?: boolean
     onClick?: () => void
 }
 
-export const Menu: React.FC<MenuProps> = ({ className, onClick }) => {
+export const Menu: React.FC<MenuProps> = ({
+    className,
+    sidebarMenu,
+    onClick
+}) => {
     const router = useRouter()
     const { t } = useTranslation()
+
+    const [dropdownOpen, setDropdownOpen] = useState<number>()
 
     const menuItems: MenuItemType[] = [
         {
@@ -32,11 +40,35 @@ export const Menu: React.FC<MenuProps> = ({ className, onClick }) => {
         },
         {
             link: '/stargazing',
-            text: t('stargazing')
+            text: t('stargazing'),
+            subMenuItems: [
+                {
+                    link: '/stargazing/rules',
+                    text: t('stargazing-rules')
+                },
+                {
+                    link: '/stargazing/howto',
+                    text: t('stargazing-howto')
+                },
+                {
+                    link: '/stargazing/where',
+                    text: t('stargazing-where')
+                },
+                {
+                    link: '/stargazing/faq',
+                    text: t('stargazing-faq')
+                }
+            ]
         },
         {
             link: '/observatory',
-            text: t('observatory')
+            text: t('observatory'),
+            subMenuItems: [
+                {
+                    link: '/observatory/overview',
+                    text: t('observatory-orenburg')
+                }
+            ]
         },
         {
             link: '/starmap',
@@ -47,17 +79,35 @@ export const Menu: React.FC<MenuProps> = ({ className, onClick }) => {
             text: t('about')
         },
         {
-            link: 'https://t.me/nearspace',
+            link: 'https://t.me/look_at_stars',
             text: t('telegram')
         }
     ]
 
+    const handleMouseEnter = (i: number) => {
+        setDropdownOpen(i)
+    }
+
+    const handleMouseLeave = () => {
+        setDropdownOpen(undefined)
+    }
+
     return (
-        <menu className={cn(className, styles.menu)}>
+        <menu
+            className={cn(
+                className,
+                styles.menu,
+                sidebarMenu && styles.sidebarMenu
+            )}
+        >
             {menuItems
                 .filter(({ link }) => !!link)
                 .map((item, i) => (
-                    <li key={`menu${i}`}>
+                    <li
+                        key={`menu${i}`}
+                        onMouseEnter={() => handleMouseEnter(i)}
+                        onMouseLeave={handleMouseLeave}
+                    >
                         <Link
                             className={
                                 router.asPath === item.link
@@ -70,7 +120,49 @@ export const Menu: React.FC<MenuProps> = ({ className, onClick }) => {
                             target={item.external ? '_blank' : undefined}
                         >
                             {item.text}
+                            {item.subMenuItems && !sidebarMenu ? (
+                                dropdownOpen === i ? (
+                                    <Icon
+                                        name={'KeyboardUp'}
+                                        className={styles.arrow}
+                                    />
+                                ) : (
+                                    <Icon
+                                        name={'KeyboardDown'}
+                                        className={styles.arrow}
+                                    />
+                                )
+                            ) : (
+                                ''
+                            )}
                         </Link>
+                        {item.subMenuItems &&
+                            (dropdownOpen === i || sidebarMenu) && (
+                                <ul className={styles.dropdownMenu}>
+                                    {item.subMenuItems.map((subItem, j) => (
+                                        <li key={`submenu${j}`}>
+                                            <Link
+                                                className={
+                                                    router.asPath ===
+                                                    subItem.link
+                                                        ? styles.active
+                                                        : undefined
+                                                }
+                                                href={subItem.link}
+                                                title={subItem.text}
+                                                onClick={() => onClick?.()}
+                                                target={
+                                                    subItem.external
+                                                        ? '_blank'
+                                                        : undefined
+                                                }
+                                            >
+                                                {subItem.text}
+                                            </Link>
+                                        </li>
+                                    ))}
+                                </ul>
+                            )}
                     </li>
                 ))}
         </menu>